relocate primitives
